Update polynomial_class.cpp
This commit so far seems to resolve the problem of two nodes being created for a single term. The problem was that once the first term was created, the program would encounter a ' ' character in the polynomial cstring on the second iteration of the while loop. On the second iteration, it would not execute any of the conditionals since it does not satisfy their conditions. After skipping the conditions, it would just store the same coefficient and exponent as the first term.